--- Content.Server/Ghost/GhostBooEvent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server.Ghost
-{
-    public sealed class GhostBooEvent : HandledEntityEventArgs
-    {
-
-    }
-}
