+++ Content.Shared/Ghost/GhostBooEvent.cs
@@ -1,0 +1,4 @@
+namespace Content.Shared.Ghost;
+
+[ByRefEvent]
+public record struct GhostBooEvent(bool Handled = false);
